+++ frontend/src/app/admin/milestone/register/page.tsx
@@ -10,7 +10,7 @@ import { useRouter } from 'next/navigation';
 import { toast } from 'react-toastify';
 import * as Yup from 'yup';
 
-import { FileUploader } from '@/app/components/Formik/FileUploader';
+import { FileUploader } from '@/components/Formik/FileUploader';
 import { useRegisterHistoryInBatchMutation } from '@/lib/hooks/useAdminApi';
 
 const validationSchema = Yup.object().shape({
