--- frontend/src/app/components/Milestone/styled.ts
@@ -1,11 +1,0 @@
-'use client';
-
-import styled from 'styled-components';
-
-export const MilestoneChartWrapper = styled.div`
-  margin: 12px;
-  display: flex;
-  gap: 20px;
-  align-items: center;
-  justify-content: space-around;
-`;
